Added bonded interface support to system stats plugin (#6668)
During the stats check this will also check the current interface directory's entries for anything titled
'slave_'.  For each one of these that exists it's entry will be added to that interface as bondname.slave_inf.speed

This will allow users to determine the actual composition of their bonded interface by parsing any slave entries off of
and interface name.  A quick check via the speed will let them know the makeup of the speed value for the bond.
Extra data could be added if wished, but it could also just be parsed from the rest of the stats since the
interface stats for the slave will also be in the output